+++ zetaclient/chains/base/observer.go
@@ -1,7 +1,6 @@
 package base
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"strconv"
@@ -11,20 +10,15 @@ import (
 	lru "github.com/hashicorp/golang-lru"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 
 	"github.com/zeta-chain/node/pkg/chains"
-	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 	observertypes "github.com/zeta-chain/node/x/observer/types"
 	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
 	"github.com/zeta-chain/node/zetaclient/chains/zrepo"
 	"github.com/zeta-chain/node/zetaclient/db"
 	"github.com/zeta-chain/node/zetaclient/logs"
 	"github.com/zeta-chain/node/zetaclient/metrics"
-	"github.com/zeta-chain/node/zetaclient/mode"
 	clienttypes "github.com/zeta-chain/node/zetaclient/types"
-	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
 
 const (
@@ -45,8 +39,8 @@ type Observer struct {
 	// chainParams contains the dynamic chain parameters of the observed chain
 	chainParams observertypes.ChainParams
 
-	// zetacoreClient is the client to interact with ZetaChain
-	zetacoreClient zrepo.ZetacoreClient
+	// zetaRepo is the repository that interacts with zetacore
+	zetaRepo *zrepo.ZetaRepo
 
 	// tssSigner is the TSS signer
 	tssSigner interfaces.TSSSigner
@@ -78,21 +72,18 @@ type Observer struct {
 
 	// stop is the channel to signal the observer to stop
 	stop chan struct{}
-
-	clientMode mode.ClientMode
 }
 
 // NewObserver creates a new base observer.
 func NewObserver(
 	chain chains.Chain,
 	chainParams observertypes.ChainParams,
-	zetacoreClient zrepo.ZetacoreClient,
+	zetaRepo *zrepo.ZetaRepo,
 	tssSigner interfaces.TSSSigner,
 	blockCacheSize int,
 	ts *metrics.TelemetryServer,
 	database *db.DB,
 	logger Logger,
-	clientMode mode.ClientMode,
 ) (*Observer, error) {
 	blockCache, err := lru.New(blockCacheSize)
 	if err != nil {
@@ -102,7 +93,7 @@ func NewObserver(
 	return &Observer{
 		chain:            chain,
 		chainParams:      chainParams,
-		zetacoreClient:   zetacoreClient,
+		zetaRepo:         zetaRepo,
 		tssSigner:        tssSigner,
 		lastBlock:        0,
 		lastBlockScanned: 0,
@@ -113,7 +104,6 @@ func NewObserver(
 		mu:               &sync.Mutex{},
 		logger:           newObserverLogger(chain, logger),
 		stop:             make(chan struct{}),
-		clientMode:       clientMode,
 	}, nil
 }
 
@@ -182,9 +172,9 @@ func (ob *Observer) SetChainParams(params observertypes.ChainParams) {
 	ob.logger.Chain.Info().Any("chain_params", params).Msg("updated chain parameters")
 }
 
-// ZetacoreClient returns the zetacore client for the observer.
-func (ob *Observer) ZetacoreClient() zrepo.ZetacoreClient {
-	return ob.zetacoreClient
+// ZetaRepo returns the zrepo.ZetaRepo repository for the observer.
+func (ob *Observer) ZetaRepo() *zrepo.ZetaRepo {
+	return ob.zetaRepo
 }
 
 // TSS returns the tss signer for the observer.
@@ -243,7 +233,7 @@ func (ob *Observer) WithLastTxScanned(txHash string) *Observer {
 	defer ob.mu.Unlock()
 
 	if ob.lastTxScanned == "" {
-		ob.logger.Chain.Debug().
+		ob.logger.Chain.Info().
 			Str(logs.FieldTx, txHash).
 			Msg("initializing last scanned transaction")
 	}
@@ -400,77 +390,6 @@ func (ob *Observer) ReadLastTxScannedFromDB() (string, error) {
 		return "", err
 	}
 	return lastTx.Hash, nil
-}
-
-// PostVoteInbound posts a vote for the given vote message and returns the ballot.
-func (ob *Observer) PostVoteInbound(
-	ctx context.Context,
-	msg *crosschaintypes.MsgVoteInbound,
-	retryGasLimit uint64,
-) (string, error) {
-	const gasLimit = zetacore.PostVoteInboundGasLimit
-
-	var (
-		txHash   = msg.InboundHash
-		coinType = msg.CoinType
-	)
-
-	logger := ob.logger.Inbound.With().
-		Str(logs.FieldTx, txHash).
-		Stringer(logs.FieldCoinType, coinType).
-		Stringer("confirmation_mode", msg.ConfirmationMode).
-		Logger()
-
-	cctxIndex := msg.Digest()
-	// The cctx is created after the inbound ballot is finalized
-	// 1. if the cctx already exists, we could try voting if the ballot is present
-	// 2. if the cctx exists but the ballot does not exist, we do not need to vote
-	_, err := ob.ZetacoreClient().GetCctxByHash(ctx, cctxIndex)
-	if err == nil {
-		// The CCTX exists, and we should still vote if there is a ballot
-		_, ballotErr := ob.ZetacoreClient().GetBallotByID(ctx, cctxIndex)
-		if ballotErr != nil {
-			// Verify that there is no ballot
-			st, ok := status.FromError(ballotErr)
-			if ok && st.Code() == codes.NotFound {
-				// Query for ballot failed, the ballot does not exist and we can return
-				logger.Info().Msg("inbound detected: CCTX exists but the ballot does not")
-				return cctxIndex, nil
-			}
-		}
-	}
-
-	// make sure the message is valid to avoid unnecessary retries
-	if err := msg.ValidateBasic(); err != nil {
-		logger.Warn().Err(err).Msg("invalid inbound vote message")
-		return "", nil
-	}
-
-	// Does not vote in dry mode.
-	if ob.clientMode == mode.DryMode {
-		logger.Info().Stringer(logs.FieldMode, mode.DryMode).Msg("skipping inbound vote")
-		return "", nil
-	}
-
-	// post vote to zetacore
-	zetaHash, ballot, err := ob.ZetacoreClient().PostVoteInbound(ctx, gasLimit, retryGasLimit, msg)
-	if err != nil {
-		logger.Error().Err(err).Msg("inbound detected: error posting vote")
-		return "", err
-	}
-
-	logger = logger.With().
-		Str(logs.FieldZetaTx, zetaHash).
-		Str(logs.FieldBallotIndex, ballot).
-		Logger()
-
-	if zetaHash == "" {
-		logger.Info().Msg("inbound detected: already voted on ballot")
-	} else {
-		logger.Info().Msg("inbound detected: vote posted")
-	}
-
-	return ballot, nil
 }
 
 // EnvVarLatestBlockByChain returns the environment variable for the last block by chain.