+++ zetaclient/chains/bitcoin/observer/outbound.go
@@ -26,19 +26,18 @@ const (
 )
 
 func (ob *Observer) ProcessOutboundTrackers(ctx context.Context) error {
-	chainID := ob.Chain().ChainId
-	trackers, err := ob.ZetacoreClient().GetOutboundTrackers(ctx, chainID)
+	trackers, err := ob.ZetaRepo().GetOutboundTrackers(ctx)
 	if err != nil {
-		return errors.Wrap(err, "unable to get all outbound trackers")
+		return err
 	}
 
 	logger := ob.logger.Outbound
 
 	for _, tracker := range trackers {
 		// get the CCTX
-		cctx, err := ob.ZetacoreClient().GetCctxByNonce(ctx, chainID, tracker.Nonce)
+		cctx, err := ob.ZetaRepo().GetCCTX(ctx, tracker.Nonce)
 		if err != nil {
-			logger.Err(err).Uint64(logs.FieldNonce, tracker.Nonce).Msg("cannot find CCTX")
+			logger.Err(err).Uint64(logs.FieldNonce, tracker.Nonce).Send()
 			break
 		}
 		if len(tracker.HashList) > 1 {
@@ -175,10 +174,8 @@ func (ob *Observer) VoteOutboundIfConfirmed(ctx context.Context, cctx *crosschai
 		receiveStatus = chains.ReceiveStatus_failed
 	}
 
-	signer := ob.ZetacoreClient().GetKeys().GetOperatorAddress()
-
 	msg := crosschaintypes.NewMsgVoteOutbound(
-		signer.String(),
+		ob.ZetaRepo().GetOperatorAddress(),
 		cctx.Index,
 		res.TxID,
 		// #nosec G115 always positive
@@ -195,22 +192,10 @@ func (ob *Observer) VoteOutboundIfConfirmed(ctx context.Context, cctx *crosschai
 		crosschaintypes.ConfirmationMode_SAFE,
 	)
 
-	zetaHash, ballot, err := ob.ZetacoreClient().PostVoteOutbound(ctx, gasLimit, gasRetryLimit, msg)
+	logger = logger.With().Str(logs.FieldTx, res.TxID).Logger()
 
-	logFields := map[string]any{
-		logs.FieldTx:          res.TxID,
-		logs.FieldZetaTx:      zetaHash,
-		logs.FieldBallotIndex: ballot,
-	}
-
-	if err != nil {
-		logger.Error().
-			Err(err).
-			Fields(logFields).
-			Msg("error confirming outbound")
-	} else if zetaHash != "" {
-		logger.Info().Fields(logFields).Msg("outbound confirmed")
-	}
+	// NOTE: ignoring VoteOutbound's errors
+	_, _, _ = ob.ZetaRepo().VoteOutbound(ctx, logger, gasLimit, gasRetryLimit, msg) //nolint:dogsled
 
 	return false, nil
 }
@@ -223,9 +208,9 @@ func (ob *Observer) refreshPendingNonce(ctx context.Context) {
 	logger := ob.logger.Outbound
 
 	// get pending nonces from zetacore
-	p, err := ob.ZetacoreClient().GetPendingNoncesByChain(ctx, ob.Chain().ChainId)
+	p, err := ob.ZetaRepo().GetPendingNonces(ctx)
 	if err != nil {
-		logger.Error().Err(err).Msg("error getting pending nonces")
+		logger.Error().Err(err).Send()
 	}
 
 	// increase pending nonce if lagged behind
@@ -247,9 +232,9 @@ func (ob *Observer) getOutboundHashByNonce(ctx context.Context, nonce uint64) (s
 		return res.TxID, nil
 	}
 
-	send, err := ob.ZetacoreClient().GetCctxByNonce(ctx, ob.Chain().ChainId, nonce)
+	send, err := ob.ZetaRepo().GetCCTX(ctx, nonce)
 	if err != nil {
-		return "", errors.Wrapf(err, "error getting cctx for nonce %d", nonce)
+		return "", err
 	}
 
 	txid := send.GetCurrentOutboundParam().Hash