+++ zetaclient/chains/evm/observer/outbound.go
@@ -29,9 +29,9 @@ import (
 
 // ProcessOutboundTrackers processes outbound trackers
 func (ob *Observer) ProcessOutboundTrackers(ctx context.Context) error {
-	trackers, err := ob.ZetacoreClient().GetOutboundTrackers(ctx, ob.Chain().ChainId)
+	trackers, err := ob.ZetaRepo().GetOutboundTrackers(ctx)
 	if err != nil {
-		return errors.Wrap(err, "GetOutboundTrackers error")
+		return err
 	}
 
 	// keep last block up-to-date
@@ -98,10 +98,10 @@ func (ob *Observer) postVoteOutbound(
 ) {
 	chainID := ob.Chain().ChainId
 
-	signerAddress := ob.ZetacoreClient().GetKeys().GetOperatorAddress()
+	signerAddress := ob.ZetaRepo().GetOperatorAddress()
 
 	msg := crosschaintypes.NewMsgVoteOutbound(
-		signerAddress.String(),
+		signerAddress,
 		cctxIndex,
 		receipt.TxHash.Hex(),
 		receipt.BlockNumber.Uint64(),
@@ -128,20 +128,8 @@ func (ob *Observer) postVoteOutbound(
 		Stringer(logs.FieldTx, receipt.TxHash).
 		Logger()
 
-	// post vote to zetacore
-	zetaTxHash, ballot, err := ob.ZetacoreClient().PostVoteOutbound(ctx, gasLimit, retryGasLimit, msg)
-	if err != nil {
-		logger.Error().Err(err).Msg("unable to post outbound vote")
-		return
-	}
-
-	// print vote tx hash and ballot
-	if zetaTxHash != "" {
-		logger.Info().
-			Str(logs.FieldZetaTx, zetaTxHash).
-			Str(logs.FieldBallotIndex, ballot).
-			Msg("posted outbound vote")
-	}
+	// NOTE: ignoring VoteOutbound's errors
+	_, _, _ = ob.ZetaRepo().VoteOutbound(ctx, logger, gasLimit, retryGasLimit, msg) //nolint:dogsled
 }
 
 // VoteOutboundIfConfirmed checks outbound status and returns (continueKeysign, error)