+++ zetaclient/chains/evm/signer/signer.go
@@ -240,7 +240,7 @@ func (signer *Signer) broadcast(ctx context.Context, tx *eth.Transaction) error 
 func (signer *Signer) TryProcessOutbound(
 	ctx context.Context,
 	cctx *crosschaintypes.CrossChainTx,
-	zetacoreClient zrepo.ZetacoreClient,
+	zetaRepo *zrepo.ZetaRepo,
 	height uint64,
 ) {
 	outboundID := base.OutboundIDFromCCTX(cctx)
@@ -262,14 +262,14 @@ func (signer *Signer) TryProcessOutbound(
 	// prepare logger and a few local variables
 	var (
 		params = cctx.GetCurrentOutboundParam()
-		myID   = zetacoreClient.GetKeys().GetOperatorAddress()
+		myID   = zetaRepo.GetOperatorAddress()
 		logger = signer.Logger().Std.With().
 			Int64(logs.FieldChain, signer.Chain().ChainId).
 			Uint64(logs.FieldNonce, params.TssNonce).
 			Str(logs.FieldCctxIndex, cctx.Index).
 			Str("cctx_receiver", params.Receiver).
 			Stringer("cctx_amount", params.Amount).
-			Stringer("signer", myID).
+			Str("signer", myID).
 			Logger()
 	)
 	logger.Info().Msg("TryProcessOutbound")
@@ -316,7 +316,7 @@ func (signer *Signer) TryProcessOutbound(
 		logger,
 		cctx,
 		txData,
-		zetacoreClient,
+		zetaRepo,
 		toChain,
 	)
 	if err != nil {
@@ -328,7 +328,7 @@ func (signer *Signer) TryProcessOutbound(
 	logger = logger.With().Str(logs.FieldTx, tx.Hash().Hex()).Logger()
 
 	// Broadcast Signed Tx
-	signer.BroadcastOutbound(ctx, tx, cctx, logger, zetacoreClient, txData)
+	signer.BroadcastOutbound(ctx, tx, cctx, logger, zetaRepo, txData)
 }
 
 // SignOutboundFromCCTX signs an outbound transaction from a given cctx
@@ -339,7 +339,7 @@ func (signer *Signer) SignOutboundFromCCTX(
 	logger zerolog.Logger,
 	cctx *crosschaintypes.CrossChainTx,
 	outboundData *OutboundData,
-	zetacoreClient zrepo.ZetacoreClient,
+	zetaRepo *zrepo.ZetaRepo,
 	_ zctx.Chain,
 ) (*eth.Transaction, error) {
 	switch {
@@ -366,7 +366,7 @@ func (signer *Signer) SignOutboundFromCCTX(
 	case cctx.ProtocolContractVersion == crosschaintypes.ProtocolContractVersion_V2:
 		// call sign outbound from cctx for v2 protocol contracts
 		return signer.SignOutboundFromCCTXV2(ctx, cctx, outboundData)
-	case IsPendingOutboundFromZetaChain(cctx, zetacoreClient):
+	case IsPendingOutboundFromZetaChain(cctx, zetaRepo):
 		switch cctx.InboundParams.CoinType {
 		case coin.CoinType_Gas:
 			logger.Info().Msg("calling SignGasWithdraw")
@@ -378,7 +378,7 @@ func (signer *Signer) SignOutboundFromCCTX(
 			logger.Info().Msg("calling SignConnectorOnReceive")
 			return signer.SignConnectorOnReceive(ctx, outboundData)
 		}
-	case cctx.CctxStatus.Status == crosschaintypes.CctxStatus_PendingRevert && cctx.OutboundParams[0].ReceiverChainId == zetacoreClient.Chain().ChainId:
+	case cctx.CctxStatus.Status == crosschaintypes.CctxStatus_PendingRevert && cctx.OutboundParams[0].ReceiverChainId == zetaRepo.ZetaChain().ChainId:
 		switch cctx.InboundParams.CoinType {
 		case coin.CoinType_Zeta:
 			logger.Info().Msg("calling SignConnectorOnRevert")
@@ -411,7 +411,7 @@ func (signer *Signer) BroadcastOutbound(
 	tx *eth.Transaction,
 	cctx *crosschaintypes.CrossChainTx,
 	logger zerolog.Logger,
-	zetacoreClient zrepo.ZetacoreClient,
+	zetaRepo *zrepo.ZetaRepo,
 	txData *OutboundData,
 ) {
 	app, err := zctx.FromContext(ctx)
@@ -465,14 +465,14 @@ func (signer *Signer) BroadcastOutbound(
 
 		// broadcast success, report to tracker
 		if err = signer.broadcast(ctx, tx); err == nil {
-			signer.reportToOutboundTracker(ctx, zetacoreClient, toChain.ID(), nonce, outboundHash, logger)
+			signer.reportToOutboundTracker(ctx, zetaRepo, toChain.ID(), nonce, outboundHash, logger)
 			return nil
 		}
 
 		// handle different broadcast errors
 		retry, report := zetacore.HandleBroadcastError(err, nonce, toChain.ID(), outboundHash)
 		if report {
-			signer.reportToOutboundTracker(ctx, zetacoreClient, toChain.ID(), nonce, outboundHash, logger)
+			signer.reportToOutboundTracker(ctx, zetaRepo, toChain.ID(), nonce, outboundHash, logger)
 			return nil
 		}
 		if retry {
@@ -498,10 +498,9 @@ func (signer *Signer) BroadcastOutbound(
 // TODO(revamp): move to another package more general for cctx functions
 func IsPendingOutboundFromZetaChain(
 	cctx *crosschaintypes.CrossChainTx,
-
-	zetacoreClient zrepo.ZetacoreClient,
+	zetaRepo *zrepo.ZetaRepo,
 ) bool {
-	return cctx.InboundParams.SenderChainId == zetacoreClient.Chain().ChainId &&
+	return cctx.InboundParams.SenderChainId == zetaRepo.ZetaChain().ChainId &&
 		cctx.CctxStatus.Status == crosschaintypes.CctxStatus_PendingOutbound
 }